86 - experimenting with future support for long lines of tokens
We need some way to wrap long lines without triggering paren-insertion.

Firstly we can ignore lines if there's a non-start-of-line paren currently
on the stack. This does for arglists like:
  def foo(a b c
          d e f)
    body

Secondly, lines with a single trailing extra space over the previous line
such as:
  (a b c
   d e f)
should be left as-is if there's an open paren.

But agh, this commit's thinking about this at the wrong level.
Sigh, we're going to need to pass yet more information from the tokenizer.