586 - no we still need elambda because macro params can hide the caller's params
So we have two different kinds of special-cases now, for primFuncs and for
elambda. Just a matter of time before I discover a macro use-case that needs
to poke at caller's lexical scope, and then ilambda will have to return as
well.

The fig leaf is that inlining alone doesn't need two separate special-cases
for primFuncs and ilambda.