784 - another hard-to-find bug. Reordering keyword args adds nil for each param, but be careful not to turn rest args from nil into (nil).
Took me a couple of hours to realize it was inserting nils into the arg list,
a couple of days off and on to realize that was a problem for varargs, and a
couple more hours to remember the quoted-varargs case. Now 'or' works right
again.

I'm just glad it wasn't another premature-GC bug.