786 - simplify interactive tokenize; always require an empty line to break off
I want to be able to spread tests out on 3 lines, but I've discovered a bug.
It's going to be hard to fix because we have no access to the indent of
previous lines. This is going to require some rearchitecting. There's going
to be broken tests for a few commits.