1059 - eliminate the coupling between cases of '='
I was resistant to this because new types won't show up as type list. But
assigning to naked objects is unlikely as well. We assign to forms, which are
always lists.

I feel better about using assignment for code organization. It's a very
specific kind of mutation, and there is a principle to my use of it. It's only
going to be a tiny fraction of a codebase. We can entangle behavior and time
for this tiny slice of code that lives in 'code space' before we return to the
problem space itself and try to act in more referentially-transparent fashion.