1566 - splicing args no longer inserts ''
Agh, this is going to need extensive surgery. Is being able to quote params
worth all this trouble? http://arclanguage.org/item?id=15813 all over again.

I currently only use selective quoting in primFuncs. On the other hand, it's
an onion if you can quote the entire param list but not individual params.

Just build it, then we'll see how to reorganize the implementation.
We'll have to move away from the splice-reorder-eval separation.