1563 - restore the splice+macro warning yet again
Even with Pauan's recommendation (http://arclanguage.org/item?id=15771),
it seems some macros still have trouble with splice because of the implicit
eval.

  (mac foo(a b)
    `(cons ,a ,b))

  (apply foo '(1 (2 3)))
  => (eval (cons foo '(1 (2 3))))
  => (eval (list foo '1 '(2 3)))
  => (eval `(,foo 1 (2 3)))
  => (foo 1 (2 3))
  => (eval `(cons 1 (2 3)))
  => error

---

The warning triggers whenever params are all quoted, will cause false
positives.

The tests for spliceArgs were incorrectly phrased; I was passing in a function
when a scope was expected.