1686 - uniformly prefer type predicates over isa
Is it redundant to have both?
The combination of '?.' is weird to the eye, especially since '?' is part
of the previous token but '.' is not. But I use cons?.x and cons?:car.x
all over the place. And if we have a predicate for cons we might as well
make it available for all types.