1914 - experiment: ':' as a comment token
Use it to help guide the eye in complex expressions.

  (if !user
      : (submit-login-warning url title showtext text)
     (~and (or blank.url valid-url.url)
           ~blank.title)
      : (submit-page user url title showtext text retry*)
     (len> title title-limit*)
      : (submit-page user url title showtext text toolong*)
     (and blank.url blank.text)
      : (let dummy 34
          (submit-page user url title showtext text bothblank*))
     (let site sitename.url
       (or big-spamsites*.site recent-spam.site))
      : (msgpage user spammage*)
     (oversubmitting user ip 'story url)
      : (msgpage user toofast*)
     : (let s (create-story url process-title.title text user ip)
         (story-ban-test user s ip url)
         (when ignored.user (kill s 'ignored))
         (submit-item user s)
         (maybe-ban-ip s)
         "newest"))

Here the colons mark branches in a complex 'if' expression, and thereby make test
expressions (even the long spam test above) stand out in relief.

A colon doesn't change the number of words on a line. This line will not trigger
paren-insertion:

  : x

It *does* affect the indentation of a line, however.

Does the leading colon look too much like a comment?