2290 - switch to scheme-style definitions
  def foo(x) => def (foo x)

Over the past few months I've been gradually realizing the reason
scheme's approach is valuable: it visually shows how a call translates
to code.

Infix makes this even more stark: def (a = b) ...

I've also been starting to grow aware of the power of
paren-before-function in simplifying macro definition. The notation f(x)
needs to die, starting with myself.

There's still some loose ends. defcall, defset, defcoerce are unchanged;
that might not be ideal.

One drawback of this change: you lose the ability to quote all params at
once.

---

In the process of this change I had to fix a subtle bug in
construct_macro_call, the helper for generics. I'm not sure how to
encode that bug in a test.