2399 - experimenting with a new kind of comment
Comments are versatile and can be landmarks or background hum:
  http://akkartik.name/blog/2012-11-24-18-10-36-soc

But I'm not sure this is a good idea. Perhaps the dichotomy happens
at a codebase level, not on a comment-by-comment basis.
Landmarks are often section headings, but not always.
It's not clear how salient they should be, and how 'dark' background comments
should be.
Perhaps this is angels and pinheads, not worth over-thinking.