2686 - expand just one level at a time
But be careful of javascript operations that need to traverse the entire
file. We want to encourage logs to grow huge.

One way to eliminate the initial 'onload' overhead is to explicitly
encode display state and onclick handlers in each line. But that'll blow
up file size.. Let's revisit this when we get to a large logfile.