2806 - subsetting multiple layers at once
I'm deliberately not writing tests for 'split'. Minimize interfaces with
tests; the more there are, the fewer the possibilities for redrawing
function boundaries.

But we can't just replace readAll with run in all tests; when testing
earlier phases it's nice to be able to provide syms without bindings.

Besides, read vs eval is a hallowed interface boundary. Unlikely we'll
think of even a plausible-seeming twist on it.