2847 - The great rename of 2013
Now we should be using the same naming conventions in both C and wart
code: underscores for everything, globals capitalized, macros and
constants in all-uppercase.

The hope is that the type-vs-variable divide is not that important, and
that highlighting globals like in lisp allows us to use them more
tastefully rather than ban them outright.