2990 - avoid peek() in interactive mode
I'd added it in 2761, back when we were checking the entire trace.

The move away from at_start_of_line also happened somewhere in the
course of the layers voyage. I took it out with parenthesize, but forgot
to bring it back.

Conundrum: how can I check in a test that the repl doesn't call peek()?