3002
Basically, read_all needs to keep going until there's no at_start_of_line
state to track. But that's true on either side of the newline, and it's
simpler for other passes to be indisciplined about consuming the
newline.

Indeed, it's easier to always leave the newline, than it is to always
consume it. What if there's multiple exprs on a line? No intervening
token to consume then. And putback is non-trivial to build for
indent_sensitive_stream.

Incidentally, tracing tokenize to screen shows a huge lag before it
prints the result. Is that part of the reason for the slowdown?