3105 - reorganize files
The 00* files are foundational for my method, useful in any domain, not
just this specific program.

Files in prefixes 1*-2* are collectively layer 0, the core skeleton for
this specific domain of lisp interpreters. We don't expect to ever take
any sub-features out.

1* is a partial walk-through of the lisp evaluation model from the
top-level down.
2* shows the rest, a walk-through of how cons cells are constructed in
the read phase.

Past them, we expect each file to correspond to a layer.

I spent a long time thinking about how to organize the files in my new
organization, but I've subsided to just this coarse division in the
early prefixes, and an ad hoc ordering for most later layers in the
absence of dependencies. Especially in the early files, a clear
narrative flow is the top priority.

We *could* order later layers any way we want even in the presence of
dependencies, but use the ordering directives only for ordering within
each layer. Try to define layers after their dependencies.

I might try to record inside later layers or scenarios any dependencies
they rely on. We'll see how that goes.