3106
I'd been hoping that the various levels of build could operate as a
visualization of the parse tree, but that's not really working out.
So go back to keeping tests in the layer they're testing, and we have
to come up with some alternative way to help visualize cons trees.