3335 - bugfix
After all that, update_optimized_body was again failing on '+'!

This fix adds two more blemishes to my inventory of 3331: the new
version is no longer shorter, and there's now a second correct_args
temporary.

I'm still assuming that the param is a rest param if the args are a
cons. Can I get away with that?